requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
altex==0.2.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.44
botocore==1.42.44
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cssselect==1.3.0
cycler==0.12.1
datasets==4.5.0
deep-translator==1.11.4
dill==0.4.0
easyocr==1.7.2
entrypoints==0.4
Faker==40.4.0
favicon==0.7.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==0.36.0
hyperframe==5.2.0
idna==2.10
ImageIO==2.37.2
jieba3k==0.35.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langdetect==1.0.9
lazy_loader==0.4
lxml==6.0.2
lxml_html_clean==0.4.3
Markdown==3.10.1
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==2.15.0
networkx==3.6.1
newspaper3k==0.2.8
ninja==1.13.0
nltk==3.9.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.4
pyarrow==22.0.0
pyclipper==1.4.0
pycparser==3.0
pydeck==0.9.1
PyJWT==2.11.0
pymdown-extensions==10.20.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.9
requests==2.32.5
requests-file==3.0.1
rfc3986==1.5.0
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
sentence-transformers==5.2.3
sgmllib3k==1.0.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.2.0
snowflake-snowpark-python==1.45.0
sortedcontainers==2.4.0
soupsieve==2.8.1
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.52.2
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.12.20
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
torch==2.9.1+cpu
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
validators==0.35.0
watchdog==6.0.0
xxhash==3.6.0
yarl==1.22.0
